Added a callback method to FileBrowserListener.

+++ modules/juce_gui_basics/keyboard/juce_TextEditorKeyMapper.h
@@ -43,59 +43,33 @@ struct TextEditorKeyMapper
     */
     static bool invokeKeyFunction (CallbackClass& target, const KeyPress& key)
     {
-        const bool isShiftDown = key.getModifiers().isShiftDown();
+        const bool isShiftDown   = key.getModifiers().isShiftDown();
         const bool ctrlOrAltDown = key.getModifiers().isCtrlDown() || key.getModifiers().isAltDown();
 
-        if (key == KeyPress (KeyPress::downKey, ModifierKeys::ctrlModifier, 0)
-             && target.scrollUp())
-            return true;
-
-        if (key == KeyPress (KeyPress::upKey, ModifierKeys::ctrlModifier, 0)
-             && target.scrollDown())
-            return true;
+        if (key == KeyPress (KeyPress::downKey, ModifierKeys::ctrlModifier, 0) && target.scrollUp())   return true;
+        if (key == KeyPress (KeyPress::upKey,   ModifierKeys::ctrlModifier, 0) && target.scrollDown()) return true;
 
        #if JUCE_MAC
         if (key.getModifiers().isCommandDown())
         {
-            if (key.isKeyCode (KeyPress::upKey))
-                return target.moveCaretToTop (isShiftDown);
-
-            if (key.isKeyCode (KeyPress::downKey))
-                return target.moveCaretToEnd (isShiftDown);
-
-            if (key.isKeyCode (KeyPress::leftKey))
-                return target.moveCaretToStartOfLine (isShiftDown);
-
-            if (key.isKeyCode (KeyPress::rightKey))
-                return target.moveCaretToEndOfLine (isShiftDown);
+            if (key.isKeyCode (KeyPress::upKey))        return target.moveCaretToTop (isShiftDown);
+            if (key.isKeyCode (KeyPress::downKey))      return target.moveCaretToEnd (isShiftDown);
+            if (key.isKeyCode (KeyPress::leftKey))      return target.moveCaretToStartOfLine (isShiftDown);
+            if (key.isKeyCode (KeyPress::rightKey))     return target.moveCaretToEndOfLine (isShiftDown);
         }
        #endif
 
-        if (key.isKeyCode (KeyPress::upKey))
-            return target.moveCaretUp (isShiftDown);
+        if (key.isKeyCode (KeyPress::upKey))        return target.moveCaretUp (isShiftDown);
+        if (key.isKeyCode (KeyPress::downKey))      return target.moveCaretDown (isShiftDown);
+        if (key.isKeyCode (KeyPress::leftKey))      return target.moveCaretLeft (ctrlOrAltDown, isShiftDown);
+        if (key.isKeyCode (KeyPress::rightKey))     return target.moveCaretRight (ctrlOrAltDown, isShiftDown);
+        if (key.isKeyCode (KeyPress::pageUpKey))    return target.pageUp (isShiftDown);
+        if (key.isKeyCode (KeyPress::pageDownKey))  return target.pageDown (isShiftDown);
 
-        if (key.isKeyCode (KeyPress::downKey))
-            return target.moveCaretDown (isShiftDown);
-
-        if (key.isKeyCode (KeyPress::leftKey))
-            return target.moveCaretLeft (ctrlOrAltDown, isShiftDown);
-
-        if (key.isKeyCode (KeyPress::rightKey))
-            return target.moveCaretRight (ctrlOrAltDown, isShiftDown);
-
-        if (key.isKeyCode (KeyPress::pageUpKey))
-            return target.pageUp (isShiftDown);
-
-        if (key.isKeyCode (KeyPress::pageDownKey))
-            return target.pageDown (isShiftDown);
-
-        if (key.isKeyCode (KeyPress::homeKey))
-            return ctrlOrAltDown ? target.moveCaretToTop (isShiftDown)
-                                 : target.moveCaretToStartOfLine (isShiftDown);
-
-        if (key.isKeyCode (KeyPress::endKey))
-            return ctrlOrAltDown ? target.moveCaretToEnd (isShiftDown)
-                                 : target.moveCaretToEndOfLine (isShiftDown);
+        if (key.isKeyCode (KeyPress::homeKey))  return ctrlOrAltDown ? target.moveCaretToTop (isShiftDown)
+                                                                     : target.moveCaretToStartOfLine (isShiftDown);
+        if (key.isKeyCode (KeyPress::endKey))   return ctrlOrAltDown ? target.moveCaretToEnd (isShiftDown)
+                                                                     : target.moveCaretToEndOfLine (isShiftDown);
 
         if (key == KeyPress ('c', ModifierKeys::commandModifier, 0)
               || key == KeyPress (KeyPress::insertKey, ModifierKeys::ctrlModifier, 0))
@@ -109,11 +83,8 @@ struct TextEditorKeyMapper
               || key == KeyPress (KeyPress::insertKey, ModifierKeys::shiftModifier, 0))
             return target.pasteFromClipboard();
 
-        if (key.isKeyCode (KeyPress::backspaceKey))
-            return target.deleteBackwards (ctrlOrAltDown);
-
-        if (key.isKeyCode (KeyPress::deleteKey))
-            return target.deleteForwards (ctrlOrAltDown);
+        if (key.isKeyCode (KeyPress::backspaceKey))     return target.deleteBackwards (ctrlOrAltDown);
+        if (key.isKeyCode (KeyPress::deleteKey))        return target.deleteForwards (ctrlOrAltDown);
 
         if (key == KeyPress ('a', ModifierKeys::commandModifier, 0))
             return target.selectAll();
